18933dd7 COMP: Resolve warning for converting to int from double
c02557d2 ENH: Add GetParticleName to internal flash reader
3e456945 ENH: Remove vtkPlane from class ivars
5bfba542 ENH: Use vtkUnstructuredGrid in CutAMRBlock
2c911dc4 ENH: Implement extraction of cells on plane
9ee40558 COMP: Fix compiler warnings
c4941ad9 ENH: Use octree point locator to extract cells
e30124f5 ENH: Remove vtkContourValues and use endl
b5a58633 ENH: Use vtkOverlappingAMR + style changes

+++ AMR/vtkAMRSliceFilter.cxx
@@ -462,7 +462,7 @@ int vtkAMRSliceFilter::GetDonorCellIdx( double x[3], vtkUniformGrid *ug )
   int ijk[3];
   for( int i=0; i < 3; ++i )
     {
-    ijk[ i ] = floor( (x[i]-x0[i])/h[i] );
+    ijk[ i ] = static_cast<int>(floor( (x[i]-x0[i])/h[i] ));
     }
 
   int dims[3];
